Change file structure for persistance
